move wsl' util_windows to windows package so it can be reused by hyperv
